Q4_work/Learn_MCP/personal_mcp_server/requirements.txt
mcp>=0.1.1
openai-agents>=0.2.5
python-dotenv>=1.0.1
rich>=13.7.1
requests>=2.32.3
uvicorn>=0.30.0